Move trait ToTokens impl to codegen module
